+++ kai/ukernels/matmul/kai_matmul.h
@@ -118,6 +118,22 @@ struct kai_matmul_uker_api kai_matmul_clamp_f32_u8p4vsx4_u8p4vsx4_i32_i32_f32_f3
 /// @return The micro-kernel API.
 struct kai_matmul_uker_api kai_matmul_clamp_f32_f32_f32p4vsx1bf32_1x32vs_sme2_mla(void);
 
+/// Single-precision floating-point vector-matrix multiplication using QMX MLA instruction.
+///
+/// Required operands:
+///   * dst
+///   * lhs
+///   * rhs - rhs matrix and per-n accumulator bias vector.
+///
+/// Optional arguments:
+///   * clamp - F32 output clamp values if KAI_MATMUL_UKER_FLAGS_ARGS_CLAMP flag is set.
+///
+/// Supported flags:
+///   * KAI_MATMUL_UKER_FLAGS_ARGS_CLAMP - Clamp output data.
+///
+/// @return The micro-kernel API.
+struct kai_matmul_uker_api kai_matmul_clamp_f32_f32_f32p4vsx1bf32_1x32vs_qmx_mla(void);
+
 /// Statically quantized INT8 matrix multiplication using SME2 MOPA instruction.
 ///
 /// Required CPU features: