+++ kai/ukernels/matmul/matmul_clamp_f32_f32_f32p/kai_matmul_clamp_f32_f32_f32p4vsx1bf32_1x32vs_qmx_mla.c
@@ -1,0 +1,195 @@
+//
+// SPDX-FileCopyrightText: Copyright 2026 Arm Limited and/or its affiliates <open-source-office@arm.com>
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+
+#if (!defined(__aarch64__) || !defined(__ARM_FEATURE_SVE2)) && !defined(_M_ARM64)
+#error This file must be compiled for AArch64, FEAT_SVE2.
+#else  // Architectural features check.
+
+#include <stddef.h>
+#include <stdint.h>
+
+#include "kai/kai_common.h"
+#include "kai/ukernels/matmul/kai_matmul.h"
+#include "kai/ukernels/matmul/kai_matmul_types.h"
+
+enum {
+    LHS_ELEM_BYTES = 4,
+    RHS_ELEM_BYTES = 4,
+    DST_ELEM_BYTES = 4,
+    ACC_ELEM_BYTES = 4,
+    BIAS_ELEM_BYTES = 4,
+
+    NR_VSCALE = 4,
+    KR = 1,
+
+    SUPPORTED_FLAGS = KAI_MATMUL_UKER_FLAGS_ARGS_CLAMP,
+};
+
+/// Matrix multiplication micro-kernel arguments.
+struct kai_matmul_uker_args_internal {
+    int32_t maxval;
+    int32_t minval;
+    const void* A_ptr;
+    const void* B_ptr;
+    size_t N;
+    size_t K;
+    void* output_ptr;
+    uint64_t flags;
+};
+
+void kai_kernel_matmul_clamp_f32_f32_f32p4vsx1bf32_1x32vs_qmx_mla(const struct kai_matmul_uker_args_internal* args);
+
+static size_t get_mr(void) {
+    return 1;
+}
+
+static size_t get_nr(void) {
+    return NR_VSCALE * kai_get_sme_vscale();
+}
+
+static size_t get_m_step(void) {
+    return get_mr();
+}
+
+static size_t get_n_step(void) {
+    return get_nr();
+}
+
+static struct kai_matmul_uker_dim_args get_step(const struct kai_matmul_uker_config* config) {
+    KAI_UNUSED(config);
+
+    const struct kai_matmul_uker_dim_args step = {
+        .m = get_m_step(),
+        .n = get_n_step(),
+        .k = 0,
+    };
+
+    return step;
+}
+
+static struct kai_matmul_uker_lhs_stride_args get_lhs_stride(
+    const struct kai_matmul_uker_config* config, const struct kai_matmul_uker_lhs_dim_args* shape) {
+    KAI_UNUSED(config);
+
+    const struct kai_matmul_uker_lhs_stride_args stride = {
+        .m = get_mr() * kai_roundup(shape->k, KR) * LHS_ELEM_BYTES,
+    };
+
+    return stride;
+}
+
+static size_t get_lhs_offset(
+    const struct kai_matmul_uker_config* config, const struct kai_matmul_uker_lhs_dim_args* index,
+    const struct kai_matmul_uker_lhs_stride_args* stride) {
+    KAI_UNUSED(config);
+    KAI_ASSUME(index->m % get_m_step() == 0);
+    KAI_ASSUME(index->k == 0);
+
+    return index->m / get_mr() * stride->m;
+}
+
+static struct kai_matmul_uker_rhs_stride_args get_rhs_stride(
+    const struct kai_matmul_uker_config* config, const struct kai_matmul_uker_rhs_dim_args* shape) {
+    KAI_UNUSED(config);
+
+    const struct kai_matmul_uker_rhs_stride_args stride = {
+        .n = get_nr() * (BIAS_ELEM_BYTES + kai_roundup(shape->k, KR) * RHS_ELEM_BYTES),
+    };
+
+    return stride;
+}
+
+static size_t get_rhs_offset(
+    const struct kai_matmul_uker_config* config, const struct kai_matmul_uker_rhs_dim_args* index,
+    const struct kai_matmul_uker_rhs_stride_args* stride) {
+    KAI_UNUSED(config);
+    KAI_ASSUME(index->n % get_n_step() == 0);
+    KAI_ASSUME(index->k == 0);
+
+    return index->n / get_nr() * stride->n;
+}
+
+static struct kai_matmul_uker_dst_stride_args get_dst_stride(
+    const struct kai_matmul_uker_config* config, const struct kai_matmul_uker_dst_dim_args* shape) {
+    KAI_UNUSED(config);
+
+    const struct kai_matmul_uker_dst_stride_args stride = {
+        .m = shape->n * DST_ELEM_BYTES,
+    };
+
+    return stride;
+}
+
+static size_t get_dst_offset(
+    const struct kai_matmul_uker_config* config, const struct kai_matmul_uker_dst_dim_args* index,
+    const struct kai_matmul_uker_dst_stride_args* stride) {
+    KAI_UNUSED(config);
+    KAI_ASSUME(index->m % get_m_step() == 0);
+    KAI_ASSUME(index->n % get_n_step() == 0);
+
+    return index->m * stride->m + index->n * DST_ELEM_BYTES;
+}
+
+static size_t get_dst_size(
+    const struct kai_matmul_uker_config* config, const struct kai_matmul_uker_dst_dim_args* shape,
+    const struct kai_matmul_uker_dst_stride_args* stride) {
+    KAI_UNUSED(config);
+
+    return shape->m * stride->m;
+}
+
+static void run(const struct kai_matmul_uker_config* config, const struct kai_matmul_uker_args* args) {
+    KAI_UNUSED(config);
+    KAI_ASSUME_MSG((args->flags & ~((size_t)SUPPORTED_FLAGS)) == 0, "Only supported flags are accepted!");
+
+    KAI_ASSUME(args->shape.m == 1);
+
+    KAI_ASSUME(args->operand.lhs.ptr != NULL);
+    KAI_ASSUME(args->operand.rhs.ptr != NULL);
+    KAI_ASSUME(args->operand.dst.ptr != NULL);
+
+    kai_commit_za();
+
+    struct kai_matmul_uker_args_internal uker_args = {
+        .maxval = 0,
+        .minval = 0,
+        .A_ptr = args->operand.lhs.ptr,
+        .B_ptr = args->operand.rhs.ptr,
+        .N = args->shape.n,
+        .K = args->shape.k,
+        .output_ptr = args->operand.dst.ptr,
+        .flags = 0,
+    };
+
+    if (args->flags & KAI_MATMUL_UKER_FLAGS_ARGS_CLAMP) {
+        uker_args.maxval = *(const int32_t*)args->activation.clamp.max_ptr;
+        uker_args.minval = *(const int32_t*)args->activation.clamp.min_ptr;
+        uker_args.flags |= 2;
+    }
+
+    kai_kernel_matmul_clamp_f32_f32_f32p4vsx1bf32_1x32vs_qmx_mla(&uker_args);
+}
+
+struct kai_matmul_uker_api kai_matmul_clamp_f32_f32_f32p4vsx1bf32_1x32vs_qmx_mla(void) {
+    struct kai_matmul_uker_api api = {
+        .run = run,
+        .get_step = get_step,
+
+        .get_lhs_stride = get_lhs_stride,
+        .get_lhs_offset = get_lhs_offset,
+
+        .get_rhs_stride = get_rhs_stride,
+        .get_rhs_offset = get_rhs_offset,
+
+        .get_dst_stride = get_dst_stride,
+        .get_dst_offset = get_dst_offset,
+        .get_dst_size = get_dst_size,
+    };
+
+    return api;
+}
+
+#endif  // Architectural features check.